Final main.js cleanup & splitted main.js to [main.js, settings,js, utils.js]
