+++ assets/style.css
@@ -1,3 +1,20 @@
+/*
+ This file is part of LVL UP Starter. LVL UP Starter is free software: you can
+ redistribute it and/or modify it under the terms of the GNU General Public
+ License as published by the Free Software Foundation, version 2.
+ 
+ This program is distributed in the hope that it will be useful, but WITHOUT
+ ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+ details.
+ 
+ You should have received a copy of the GNU General Public License along with
+ this program; if not, write to the Free Software Foundation, Inc., 51
+ Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ 
+ Copyright (C) 2015 Michał Frąckiewicz
+ */
+
 html {
     position: relative;
     min-height: 100%;
