+++ index.html
@@ -125,9 +125,11 @@
         <script type="text/javascript" src="src/locales/en.js"></script>
         <script type="text/javascript" src="src/locales/pl.js"></script>
         
-        <script type="text/javascript" src="src/download.js"></script>
         <script type="text/javascript" src="src/mc.js"></script>
         <script type="text/javascript" src="src/tools.js"></script>
+        <script type="text/javascript" src="src/utils.js"></script>
+        <script type="text/javascript" src="src/settings.js"></script>
+        <script type="text/javascript" src="src/download.js"></script>
         <script type="text/javascript" src="src/yggdrasil.js"></script>
         <script type="text/javascript" src="src/main.js"></script>
     </body>
